# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpenPlanner Local API
  version: 0.1.0
  description: |
    Spec-first contract for the local-first OpenPlanner SDK. The generated Go
    client is the primary shippable artifact in v1.
servers:
  - url: http://openplanner.local
...
paths:
  /v1/calendars:
    get:
      operationId: listCalendars
      tags: [calendars]
      summary: List calendars
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Calendar page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createCalendar
      tags: [calendars]
      summary: Create a calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarRequest'
      responses:
        '201':
          description: Created calendar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/calendars/{calendarId}:
    parameters:
      - $ref: '#/components/parameters/CalendarId'
    get:
      operationId: getCalendar
      tags: [calendars]
      summary: Get a calendar
      responses:
        '200':
          description: Calendar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateCalendar
      tags: [calendars]
      summary: Update a calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCalendarRequest'
      responses:
        '200':
          description: Updated calendar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteCalendar
      tags: [calendars]
      summary: Delete a calendar
      responses:
        '204':
          description: Calendar deleted
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List event series
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/CalendarFilter'
      responses:
        '200':
          description: Event page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event series
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Created event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/events/{eventId}:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event series
      responses:
        '200':
          description: Event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event series
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Updated event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event series
      responses:
        '204':
          description: Event deleted
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List task series
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/CalendarFilter'
      responses:
        '200':
          description: Task page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task series
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task series
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task series
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task series
      responses:
        '204':
          description: Task deleted
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/tasks/{taskId}/complete:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Complete a task occurrence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTaskRequest'
      responses:
        '200':
          description: Recorded task completion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCompletion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/agenda:
    get:
      operationId: listAgenda
      tags: [agenda]
      summary: List agenda occurrences in a range
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: Inclusive range start in RFC3339 format.
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: Exclusive range end in RFC3339 format.
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Agenda page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgendaListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque cursor returned by a previous list response.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Maximum number of items to return.
    CalendarId:
      name: calendarId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
    CalendarFilter:
      name: calendarId
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/ResourceId'
      description: Restrict results to a single calendar.
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFoundError:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ConflictError:
      description: Request conflicts with current state.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalError:
      description: Unexpected server error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Calendar:
      type: object
      additionalProperties: false
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCalendarRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
    UpdateCalendarRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
    CalendarListResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Calendar'
        nextCursor:
          $ref: '#/components/schemas/CursorToken'
    Event:
      type: object
      additionalProperties: false
      required: [id, calendarId, title, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        calendarId:
          $ref: '#/components/schemas/ResourceId'
        title:
          type: string
          minLength: 1
        description:
          type: string
        location:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateEventRequest:
      type: object
      additionalProperties: false
      required: [calendarId, title]
      properties:
        calendarId:
          $ref: '#/components/schemas/ResourceId'
        title:
          type: string
          minLength: 1
        description:
          type: string
        location:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
    UpdateEventRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        location:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
    EventListResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        nextCursor:
          $ref: '#/components/schemas/CursorToken'
    Task:
      type: object
      additionalProperties: false
      required: [id, calendarId, title, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        calendarId:
          $ref: '#/components/schemas/ResourceId'
        title:
          type: string
          minLength: 1
        description:
          type: string
        dueAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTaskRequest:
      type: object
      additionalProperties: false
      required: [calendarId, title]
      properties:
        calendarId:
          $ref: '#/components/schemas/ResourceId'
        title:
          type: string
          minLength: 1
        description:
          type: string
        dueAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
    UpdateTaskRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        dueAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
    TaskListResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        nextCursor:
          $ref: '#/components/schemas/CursorToken'
...